refactor: reorg types and update imports
- move CellRepresentation to czb.tasks.types
- move ListLike to czb.types
