+++ src/czbenchmarks/datasets/__init__.py
@@ -2,7 +2,7 @@ from .utils import load_dataset, list_available_datasets
 from .single_cell_labeled import SingleCellLabeledDataset
 from .single_cell_perturbation import SingleCellPerturbationDataset
 from .dataset import Dataset
-from .types import DataValue, Organism
+from .types import Organism
 
 __all__ = [
     "load_dataset",
@@ -10,6 +10,5 @@ __all__ = [
     "SingleCellLabeledDataset",
     "SingleCellPerturbationDataset",
     "Dataset",
-    "DataValue",
     "Organism",
 ]
