+++ src/czbenchmarks/tasks/embedding.py
@@ -2,7 +2,8 @@ import logging
 from typing import List
 
 from ..constants import RANDOM_SEED
-from ..datasets.types import CellRepresentation, ListLike
+from ..tasks.types import CellRepresentation
+from ..types import ListLike
 from ..metrics import metrics_registry
 from ..metrics.types import MetricResult, MetricType
 from .base import BaseTask
