+++ src/czbenchmarks/tasks/label_prediction.py
@@ -17,7 +17,8 @@ from sklearn.neighbors import KNeighborsClassifier
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import StandardScaler
 
-from ..datasets.types import CellRepresentation, ListLike
+from ..tasks.types import CellRepresentation
+from ..types import ListLike
 from ..metrics import metrics_registry
 from ..metrics.types import MetricResult, MetricType
 from .base import BaseTask
