+++ src/czbenchmarks/tasks/single_cell/cross_species.py
@@ -4,7 +4,8 @@ import numpy as np
 
 from ...constants import RANDOM_SEED
 from ..base import BaseTask
-from ...datasets.types import CellRepresentation, ListLike
+from ...tasks.types import CellRepresentation
+from ...types import ListLike
 from ...metrics import metrics_registry
 from ...metrics.types import MetricResult, MetricType
