+++ src/czbenchmarks/tasks/utils.py
@@ -6,7 +6,7 @@ import pandas as pd
 import scanpy as sc
 from anndata import AnnData
 from ..constants import RANDOM_SEED
-from ..datasets.types import CellRepresentation
+from ..tasks.types import CellRepresentation
 from .constants import FLAVOR, KEY_ADDED, OBSM_KEY
 
 logger = logging.getLogger(__name__)
