+++ src/czbenchmarks/types.py
@@ -1,0 +1,5 @@
+import numpy as np
+import pandas as pd
+
+
+ListLike = list | np.ndarray | pd.Series
